Make the repo more tube-like.
On every action a merge with the previous branch is done. Then it is decided
if a new branch is created or a old will be used. Six out auf ten times
an old branch is used to create a better tube effect.

Also check out a random branch if going back.